Football Endzone

The endzones are the area beyond the goal line. Any offensive player who is in legal possession of the ball while in the endzone, or who catches the ball in the endzone, scores a touchdown for their team. This area stretches from the back field boundary to the edge of the goal line closest to the center of the field

Usage

football_endzone(field_width = 0, endzone_length = 0)

Arguments

field_width

The width of the field

endzone_length

The length of the endzone

Value

A data frame of the bounding box of the endzone