Set the colors to be used for the plot. The values provided in the arguments are the defaults, and, where specified, are the rule-book specified values.

Hexadecimal values are the passed vales to this function by default, but it is also possible to use string-named values (e.g. "dodgerblue") when specifying.

Usage

football_features_set_colors(
  plot_background = NULL,
  field_apron = "#196f0c",
  offensive_half = "#196f0c",
  defensive_half = "#196f0c",
  offensive_endzone = "#196f0c",
  defensive_endzone = "#196f0c",
  end_line = "#ffffff",
  sideline = "#ffffff",
  field_border = "#196f0c",
  field_border_outline = "#ffffff00",
  red_zone_border = "#196f0c00",
  red_zone_border_outline = "#ffffff00",
  major_yard_line = "#ffffff",
  goal_line = "#ffffff",
  minor_yard_line = "#ffffff",
  directional_arrow = "#ffffff",
  try_mark = "#ffffff",
  yardage_marker = "#ffffff",
  restricted_area = "#ffffff",
  coaching_box = "#ffffff",
  team_bench_area = "#196f0c",
  team_bench_area_outline = "#ffffff",
  coaching_box_line = "#ffcb05"
)

Arguments

plot_background

A hexadecimal string representing the color to use for this feature

field_apron

A hexadecimal string representing the color to use for this feature

offensive_half

A hexadecimal string representing the color to use for this feature

defensive_half

A hexadecimal string representing the color to use for this feature

offensive_endzone

A hexadecimal string representing the color to use for this feature

defensive_endzone

A hexadecimal string representing the color to use for this feature

end_line

A hexadecimal string representing the color to use for this feature

sideline

A hexadecimal string representing the color to use for this feature

field_border

A hexadecimal string representing the color to use for this feature

field_border_outline

A hexadecimal string representing the color to use for this feature

red_zone_border

A hexadecimal string representing the color to use for this feature

red_zone_border_outline

A hexadecimal string representing the color to use for this feature

major_yard_line

A hexadecimal string representing the color to use for this feature

goal_line

A hexadecimal string representing the color to use for this feature

minor_yard_line

A hexadecimal string representing the color to use for this feature

directional_arrow

A hexadecimal string representing the color to use for this feature

try_mark

A hexadecimal string representing the color to use for this feature

yardage_marker

A hexadecimal string representing the color to use for this feature

restricted_area

A hexadecimal string representing the color to use for this feature

coaching_box

A hexadecimal string representing the color to use for this feature

team_bench_area

A hexadecimal string representing the color to use for this feature

team_bench_area_outline

A hexadecimal string representing the color to use for this feature

coaching_box_line

A hexadecimal string representing the color to use for this feature

Value

A list of hexadecimal colors to use to color the features on the resulting plot