The non-centered faceoff spots are located in the neutral, offensive and defensive zones of the ice, with one on each side of the x-axis when viewing the rink in TV view. These spots differ from the center faceoff spot because they have a larger diameter, differ in color, and have a colored stripe that runs through its center.

This function is responsible for creating the inner stripe, not the colored outer ring around it. Please see hockey_nodzone_faceoff_spot_ring() for more information on it

Usage

hockey_nodzone_faceoff_spot_stripe(
  feature_radius = 0,
  feature_thickness = 0,
  gap_width = 0
)

Arguments

feature_radius

The outer radius of the non-centered faceoff spot

feature_thickness

The thickness of the non-centered faceoff spot ring

gap_width

The width of the gap from the inner edge of the non-centered faceoff spot ring to the outer edge of the stripe

Value

A data frame containing the bounding coordinates of the non-centered faceoff spot's stripe

Details

The non-centered faceoff spots are where faceoffs are taken after an icing call or to start a powerplay. They differ from the center ice faceoff spot in size, color, and form. For the faceoff spot's stripe, the 'feature_thickness' parameter should be the thickness of the outer ring, which is 'minor_line_thickness'