Set the colors to be used for the plot. The values provided in the arguments are the defaults, and, where specified, are the rule-book specified values.

Hexadecimal values are the passed vales to this function by default, but it is also possible to use string-named values (e.g. "dodgerblue") when specifying.

Usage

lacrosse_features_set_colors(
  plot_background = NULL,
  field_apron = "#195f0c",
  defensive_zone = "#195f0c",
  neutral_zone = "#195f0c",
  offensive_zone = "#195f0c",
  team_a_bench = "#a5acaf4d",
  team_b_bench = "#a5acaf4d",
  team_a_penalty_box = "#a5acaf4d",
  team_b_penalty_box = "#a5acaf4d",
  off_field_officials_box = "#a5acaf",
  boards = "#ffa500",
  end_line = "#ffffff",
  sideline = "#ffffff",
  center_line = "#ffffff",
  wing_line = "#ffffff",
  restraining_line = "#ffffff",
  defensive_area_line = "#ffffff",
  goal_line = "#ffffff",
  referee_crease = "#ffffff",
  referee_crease_fill = "#195f0c",
  goal_circle = "#ffffff",
  goal_circle_fill = "#195f0c",
  goal_arc = "#ffffff",
  goal_fan = "#ffffff",
  goal_fan_hash_mark = "#ffffff",
  goal_mouth_hash_mark = "#ffffff",
  goal_mouth = "#ffffff",
  below_goal_marking = "#ffffff",
  goal_frame = "#ffa500",
  goal_net = "#a5acaf4d",
  center_circle = "#ffffff",
  center_face_off_marker = "#ffffff",
  corner_face_off_marker = "#ffffff",
  change_area_outline = "#ffffff",
  change_area_fill = "#195f0c"
)

Arguments

plot_background

A hexadecimal string representing the color to use for this feature

field_apron

A hexadecimal string representing the color to use for this feature

defensive_zone

A hexadecimal string representing the color to use for this feature

neutral_zone

A hexadecimal string representing the color to use for this feature

offensive_zone

A hexadecimal string representing the color to use for this feature

team_a_bench

A hexadecimal string representing the color to use for this feature

team_b_bench

A hexadecimal string representing the color to use for this feature

team_a_penalty_box

A hexadecimal string representing the color to use for this feature

team_b_penalty_box

A hexadecimal string representing the color to use for this feature

off_field_officials_box

A hexadecimal string representing the color to use for this feature

boards

A hexadecimal string representing the color to use for this feature

end_line

A hexadecimal string representing the color to use for this feature

sideline

A hexadecimal string representing the color to use for this feature

center_line

A hexadecimal string representing the color to use for this feature

wing_line

A hexadecimal string representing the color to use for this feature

restraining_line

A hexadecimal string representing the color to use for this feature

defensive_area_line

A hexadecimal string representing the color to use for this feature

goal_line

A hexadecimal string representing the color to use for this feature

referee_crease

A hexadecimal string representing the color to use for this feature

referee_crease_fill

A hexadecimal string representing the color to use for this feature

goal_circle

A hexadecimal string representing the color to use for this feature

goal_circle_fill

A hexadecimal string representing the color to use for this feature

goal_arc

A hexadecimal string representing the color to use for this feature

goal_fan

A hexadecimal string representing the color to use for this feature

goal_fan_hash_mark

A hexadecimal string representing the color to use for this feature

goal_mouth_hash_mark

A hexadecimal string representing the color to use for this feature

goal_mouth

A hexadecimal string representing the color to use for this feature

below_goal_marking

A hexadecimal string representing the color to use for this feature

goal_frame

A hexadecimal string representing the color to use for this feature

goal_net

A hexadecimal string representing the color to use for this feature

center_circle

A hexadecimal string representing the color to use for this feature

center_face_off_marker

A hexadecimal string representing the color to use for this feature

corner_face_off_marker

A hexadecimal string representing the color to use for this feature

change_area_outline

A hexadecimal string representing the color to use for this feature

change_area_fill

A hexadecimal string representing the color to use for this feature

Value

A list of hexadecimal colors to use to color the features on the resulting plot