The restraining line spans the entire width of the field (where present) and connects to the defensive area lines (see lacrosse_defensive_area_line())

The restraining line spans the entire width of the field (where present) and connects to the defensive area lines (see lacrosse_defensive_area_line())

Usage

lacrosse_restraining_line(field_width = 0, line_thickness = 0)

Arguments

field_width

The width of the field

line_thickness

The thickness of the restraining line

Value

A data frame containing the bounding coordinates of the restraining line